split into several files + start of casablanca use
